Fix bug for bedfile option.
Start and ends were defined outside the loop, they need to be input as optional arguments.

Add if case for feature definition in bedfile case. Feature is hardcoded to adapt to the bedfile case (analogous to how the bin case is handled).

Add if case for df.index.set_name in bedfile case. Analogous to how bin case is handled.